+++ niffler_e_2_e_tests_python/tests_api/enums/http_methods.py
